backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
passlib[bcrypt]
python-jose[cryptography]
python-dotenv